services/fastapi/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
# Allow newer httpx to satisfy weaviate-client (requires 0.27.0)
httpx>=0.27.0,<0.28.0
duckdb==0.10.2
langchain==0.1.16
weaviate-client==4.5.4
sentence-transformers==2.6.1
PyJWT[crypto]==2.8.0
prometheus_client==0.20.0
